+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Added enabled boolean to the messageInputBuilder and made parameters named
 - Added autoScrollTriggerOffset to the ChatPaginationControls to adjust when the auto scroll should be enabled
 - Added the ability to set the color of the CircularProgressIndicator of the ImageLoadingSnackbar by theme.snackBarTheme.actionTextColor
+- Added semantics for variable text, buttons and textfields
 
 ## 4.0.0
 - Move to the new user story architecture